Move FindSize() helper func to platform agnostic location
